fix draft handling
My approach in commit 99777d244c ("add draft label for future posts") was bogus
and showed that I just hadn't found the right docs, yet.

This is how you do it.  A filename without date component.

In the process, add a Dutch draft for the 4.9.4 announcemnt (with just the
title and the first line translated).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>